factor out util methods
also do some cleanup
